requirements.txt
numpy>=1.20.2,<1.21
pillow>=8.2.0,<8.3
pytorch>=1.11.0,<1.12
torchvision>=0.12.0,<0.13
tqdm>=4.60.0,<4.61
scipy>=1.8.0,<1.9.0
cudatoolkit>=11.3.1,<11.4
pandas>=1.2.0,<1.3.0
python>=3.8.0,<4.0.0
toml>=0.10.0,<0.11.0
